# perf: change underlying value from string to interface{}