--- a/comparison.go
+++ b/comparison.go
@@ -1,21 +1,35 @@
 package expr
 
 import (
+	"fmt"
 	"go/ast"
 	"go/token"
-	"strconv"
+
+	"github.com/muktihari/expr/conv"
 )
 
 func comparison(v, vx, vy *Visitor, binaryExpr *ast.BinaryExpr) {
 	v.kind = KindBoolean
-	if vx.kind == KindBoolean || vy.kind == KindBoolean {
-		compareBoolean(v, vx, vy, binaryExpr)
+	if vx.kind == KindBoolean && vy.kind == KindBoolean {
+		compareMustBoolean(v, vx, vy, binaryExpr)
 		return
 	}
-	if vx.kind == KindString || vy.kind == KindString {
-		compareString(v, vx, vy, binaryExpr)
+	if vx.kind == KindString && vy.kind == KindString {
+		compareMustString(v, vx, vy, binaryExpr)
 		return
 	}
+
+	// numeric can be compare one another e.g. 0.4 < 1 -> true
+	// numeric guards:
+	if vx.kind <= numeric_beg || vx.kind >= numeric_end {
+		v.err = newComparisonNonNumericError(vx, binaryExpr.X)
+		return
+	}
+	if vy.kind <= numeric_beg || vy.kind >= numeric_end {
+		v.err = newComparisonNonNumericError(vy, binaryExpr.Y)
+		return
+	}
+
 	if vx.kind == KindImag || vy.kind == KindImag {
 		compareComplex(v, vx, vy, binaryExpr)
 		return
@@ -24,16 +38,30 @@ func comparison(v, vx, vy *Visitor, binaryExpr *ast.BinaryExpr) {
 		compareFloat(v, vx, vy, binaryExpr)
 		return
 	}
-
-	compareInt(v, vx, vy, binaryExpr)
+	if vx.kind == KindInt || vy.kind == KindInt {
+		compareInt(v, vx, vy, binaryExpr)
+		return
+	}
 }
 
-func compareBoolean(v, vx, vy *Visitor, binaryExpr *ast.BinaryExpr) {
+func newComparisonNonNumericError(v *Visitor, e ast.Expr) error {
+	s := conv.FormatExpr(e)
+	return &SyntaxError{
+		Msg: "result of \"" + s + "\" is \"" + fmt.Sprintf("%v", v.value) + "\" which is not a number",
+		Pos: v.pos,
+		Err: ErrComparisonOperation,
+	}
+}
+
+func compareMustBoolean(v, vx, vy *Visitor, binaryExpr *ast.BinaryExpr) {
+	x := vx.value.(bool)
+	y := vy.value.(bool)
+
 	switch binaryExpr.Op {
 	case token.EQL:
-		v.value = strconv.FormatBool(vx.value == vy.value)
+		v.value = x == y
 	case token.NEQ:
-		v.value = strconv.FormatBool(vx.value != vy.value)
+		v.value = x != y
 	default:
 		v.kind = KindIllegal
 		v.err = &SyntaxError{
@@ -45,20 +73,23 @@ func compareBoolean(v, vx, vy *Visitor, binaryExpr *ast.BinaryExpr) {
 	}
 }
 
-func compareString(v, vx, vy *Visitor, binaryExpr *ast.BinaryExpr) {
+func compareMustString(v, vx, vy *Visitor, binaryExpr *ast.BinaryExpr) {
+	x := vx.value.(string)
+	y := vy.value.(string)
+
 	switch binaryExpr.Op {
 	case token.EQL:
-		v.value = strconv.FormatBool(vx.value == vy.value)
+		v.value = x == y
 	case token.NEQ:
-		v.value = strconv.FormatBool(vx.value != vy.value)
+		v.value = x != y
 	case token.GTR:
-		v.value = strconv.FormatBool(vx.value > vy.value)
+		v.value = x > y
 	case token.GEQ:
-		v.value = strconv.FormatBool(vx.value >= vy.value)
+		v.value = x >= y
 	case token.LSS:
-		v.value = strconv.FormatBool(vx.value < vy.value)
+		v.value = x < y
 	case token.LEQ:
-		v.value = strconv.FormatBool(vx.value <= vy.value)
+		v.value = x <= y
 	}
 }
 
@@ -69,9 +100,9 @@ func compareComplex(v, vx, vy *Visitor, binaryExpr *ast.BinaryExpr) {
 
 	switch binaryExpr.Op {
 	case token.EQL:
-		v.value = strconv.FormatBool(x == y)
+		v.value = x == y
 	case token.NEQ:
-		v.value = strconv.FormatBool(x != y)
+		v.value = x != y
 	default:
 		v.kind = KindIllegal
 		v.err = &SyntaxError{
@@ -89,36 +120,36 @@ func compareFloat(v, vx, vy *Visitor, binaryExpr *ast.BinaryExpr) {
 
 	switch binaryExpr.Op {
 	case token.EQL:
-		v.value = strconv.FormatBool(x == y)
+		v.value = x == y
 	case token.NEQ:
-		v.value = strconv.FormatBool(x != y)
+		v.value = x != y
 	case token.GTR:
-		v.value = strconv.FormatBool(x > y)
+		v.value = x > y
 	case token.GEQ:
-		v.value = strconv.FormatBool(x >= y)
+		v.value = x >= y
 	case token.LSS:
-		v.value = strconv.FormatBool(x < y)
+		v.value = x < y
 	case token.LEQ:
-		v.value = strconv.FormatBool(x <= y)
+		v.value = x <= y
 	}
 }
 
 func compareInt(v, vx, vy *Visitor, binaryExpr *ast.BinaryExpr) {
-	x, _ := strconv.ParseInt(vx.value, 0, 64)
-	y, _ := strconv.ParseInt(vy.value, 0, 64)
+	x := vx.value.(int64)
+	y := vy.value.(int64)
 
 	switch binaryExpr.Op {
 	case token.EQL:
-		v.value = strconv.FormatBool(x == y)
+		v.value = x == y
 	case token.NEQ:
-		v.value = strconv.FormatBool(x != y)
+		v.value = x != y
 	case token.GTR:
-		v.value = strconv.FormatBool(x > y)
+		v.value = x > y
 	case token.GEQ:
-		v.value = strconv.FormatBool(x >= y)
+		v.value = x >= y
 	case token.LSS:
-		v.value = strconv.FormatBool(x < y)
+		v.value = x < y
 	case token.LEQ:
-		v.value = strconv.FormatBool(x <= y)
+		v.value = x <= y
 	}
 }