--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,4 @@ module github.com/muktihari/expr
 
 go 1.13
 
-require github.com/google/go-cmp v0.5.9 // indirect
+require github.com/google/go-cmp v0.6.0