--- a/logical.go
+++ b/logical.go
@@ -1,9 +1,9 @@
 package expr
 
 import (
+	"fmt"
 	"go/ast"
 	"go/token"
-	"strconv"
 
 	"github.com/muktihari/expr/conv"
 )
@@ -19,22 +19,22 @@ func logical(v, vx, vy *Visitor, binaryExpr *ast.BinaryExpr) {
 		return
 	}
 
-	x, _ := strconv.ParseBool(vx.value)
-	y, _ := strconv.ParseBool(vy.value)
+	x := vx.value.(bool)
+	y := vy.value.(bool)
 
 	v.kind = KindBoolean
 	if binaryExpr.Op == token.LAND {
-		v.value = strconv.FormatBool(x && y)
+		v.value = x && y
 		return
 	}
 
-	v.value = strconv.FormatBool(x || y) // token.LOR
+	v.value = x || y // token.LOR
 }
 
 func newLogicalNonBooleanError(v *Visitor, e ast.Expr) error {
 	s := conv.FormatExpr(e)
 	return &SyntaxError{
-		Msg: "result of \"" + s + "\" is \"" + v.value + "\" which is not a boolean",
+		Msg: "result of \"" + s + "\" is \"" + fmt.Sprintf("%v", v.value) + "\" which is not a boolean",
 		Pos: v.pos,
 		Err: ErrLogicalOperation,
 	}