--- a/logical_test.go
+++ b/logical_test.go
@@ -12,15 +12,15 @@ func TestLogical(t *testing.T) {
 	tt := []struct {
 		v, vx, vy      *Visitor
 		ops            []token.Token
-		expectedValues []string
+		expectedValues []interface{}
 		expectedErrs   []error
 	}{
 		{
 			v:              &Visitor{},
-			vx:             &Visitor{value: "true", kind: KindBoolean},
+			vx:             &Visitor{value: true, kind: KindBoolean},
 			ops:            []token.Token{token.LAND, token.NEQ},
-			vy:             &Visitor{value: "false", kind: KindBoolean},
-			expectedValues: []string{"false", "true"},
+			vy:             &Visitor{value: false, kind: KindBoolean},
+			expectedValues: []interface{}{false, true},
 			expectedErrs:   []error{nil, nil},
 		},
 		{
@@ -28,7 +28,7 @@ func TestLogical(t *testing.T) {
 			vx:             &Visitor{value: "1", kind: KindInt},
 			ops:            []token.Token{token.LAND},
 			vy:             &Visitor{value: "false", kind: KindBoolean},
-			expectedValues: []string{""},
+			expectedValues: []interface{}{nil},
 			expectedErrs:   []error{ErrLogicalOperation},
 		},
 		{
@@ -36,7 +36,7 @@ func TestLogical(t *testing.T) {
 			vx:             &Visitor{value: "false", kind: KindBoolean},
 			ops:            []token.Token{token.LAND},
 			vy:             &Visitor{value: "1", kind: KindInt},
-			expectedValues: []string{""},
+			expectedValues: []interface{}{nil},
 			expectedErrs:   []error{ErrLogicalOperation},
 		},
 	}
@@ -49,11 +49,11 @@ func TestLogical(t *testing.T) {
 				expectedErr   = tc.expectedErrs[i]
 				expectedValue = tc.expectedValues[i]
 				be            = &ast.BinaryExpr{
-					X:  &ast.BasicLit{Value: tc.vx.value},
+					X:  &ast.BasicLit{Value: fmt.Sprintf("%v", tc.vx.value)},
 					Op: op,
-					Y:  &ast.BasicLit{Value: tc.vy.value},
+					Y:  &ast.BasicLit{Value: fmt.Sprintf("%v", tc.vy.value)},
 				}
-				name = fmt.Sprintf("%s %s %s", tc.vx.value, op, tc.vy.value)
+				name = fmt.Sprintf("%v %s %v", tc.vx.value, op, tc.vy.value)
 			)
 
 			t.Run(name, func(t *testing.T) {