… arithmetic_test.go
… comparison_test.go
… go.sum
… visitor_test.go