- Change underlying value from string to interface value, reducing unnecessary string conversion overhead.
- Use sync.Pool to handle Visitor creation to reduce allocs, instead of creating new Visitor each Visit, we can use existing Visitors. Look at this code example, after `Visitor X` finish visiting, `Visitor Y` might reuse Visitors created during previous `Visitor X`'s Visit. This will happen recursively. And also, next time user trigger any `expr` function, chances are they can ruse existing Visitor objects from previous invocation, only if it is still being kept alive by sync.Pool.
```go
func (v *Visitor) visitBinary(binaryExpr *ast.BinaryExpr) ast.Visitor {
    vx := visitorPool.Get().(*Visitor)
    defer visitorPool.Put(vx)
    vx.reset()
    vx.options = v.options

    ast.Walk(vx, binaryExpr.X)
    if vx.err != nil {
         v.err = vx.err
        return nil
    }

    /* Visitor Y might reuse existing Visitor objects created during Visitor X's Visit 
       (or from previous invocation) */
    vy := visitorPool.Get().(*Visitor)
    defer visitorPool.Put(vy)
    vy.reset()
    vy.options = v.options

    ast.Walk(vy, binaryExpr.Y)
    if vy.err != nil {
        v.err = vy.err
        return nil
    }
...
```

Here is the `benchstat` from old Version that use string with this new Version using interface plus sync.Pool.

```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/expr
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                  │    old.txt    │               xxx.txt                │
                  │    sec/op     │    sec/op     vs base                │
Any/KindBoolean-4    13.51µ ±  3%   13.28µ ± 31%        ~ (p=0.631 n=10)
Any/KindInt-4        7.053µ ± 60%   5.364µ ± 14%  -23.94% (p=0.000 n=10)
Any/KindFloat-4     11.209µ ±  4%   9.288µ ± 31%  -17.14% (p=0.029 n=10)
Any/KindImag-4       9.600µ ± 72%   5.651µ ± 17%  -41.14% (p=0.000 n=10)
geomean              10.06µ         7.819µ        -22.29%

                  │   old.txt    │               xxx.txt                │
                  │     B/op     │     B/op      vs base                │
Any/KindBoolean-4   5.297Ki ± 0%   2.734Ki ± 0%  -48.38% (p=0.000 n=10)
Any/KindInt-4       2.664Ki ± 0%   1.562Ki ± 0%  -41.35% (p=0.000 n=10)
Any/KindFloat-4     3.844Ki ± 0%   2.047Ki ± 0%  -46.75% (p=0.000 n=10)
Any/KindImag-4      2.719Ki ± 0%   1.641Ki ± 0%  -39.66% (p=0.000 n=10)
geomean             3.485Ki        1.946Ki       -44.15%

                  │   old.txt   │              xxx.txt               │
                  │  allocs/op  │ allocs/op   vs base                │
Any/KindBoolean-4   111.00 ± 0%   72.00 ± 0%  -35.14% (p=0.000 n=10)
Any/KindInt-4        62.00 ± 0%   44.00 ± 0%  -29.03% (p=0.000 n=10)
Any/KindFloat-4      91.00 ± 0%   61.00 ± 0%  -32.97% (p=0.000 n=10)
Any/KindImag-4       77.00 ± 0%   49.00 ± 0%  -36.36% (p=0.000 n=10)
geomean              83.33        55.47       -33.43%

```